+++ COREdcs/clientGUI/include/basewidget.h
@@ -26,7 +26,7 @@ class BaseWidget : public QWidget {
     virtual void startup();
     virtual void controllerInit(UA_Client *, UA_ClientConfig *,
                                 UA_CreateSubscriptionResponse) = 0;
-    virtual void updateStatus(UA_Variant data);
+    virtual void updateStatus(UA_DataValue *data);
 };
 
 #endif  // BASEWIDGET_H